+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.13.0
 - Changed query clauses `using` and `usingAll` to **not** automatically grab a `read` entitlement.  You'll need to explicitly suffix with `.read` from now on.  (`with` and `without` still automatically claim `read`.)
+- Introduced a `create` entitlement that doesn't create precedence dependencies with other `create` systems, and that will be able to run concurrently under multi-threading.
 
 ### 0.12.3
 - Added `allSystems` to the schedule builder object to facilitate sweeping precedence constraints.